+++ Combat.rtf
@@ -1,0 +1,1 @@
+{\rtf1}
