--- RPG/Combat.rtf
@@ -1,1 +1,0 @@
-{\rtf1}
