--- RPG/RPG.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29503.13
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "RPG", "RPG\RPG.vcxproj", "{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Debug|x64.ActiveCfg = Debug|x64
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Debug|x64.Build.0 = Debug|x64
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Debug|x86.ActiveCfg = Debug|Win32
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Debug|x86.Build.0 = Debug|Win32
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Release|x64.ActiveCfg = Release|x64
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Release|x64.Build.0 = Release|x64
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Release|x86.ActiveCfg = Release|Win32
-		{C7499E83-AF6C-41A2-8ACC-8D8AA0A436C5}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {D748786D-D5BA-4BED-9B12-050C5E93CB63}
-	EndGlobalSection
-EndGlobal
