--- RPG/RPG/Enemies.h
@@ -1,5 +1,0 @@
-#pragma once
-class Enemies
-{
-};
-
