--- RPG/RPG/x64/Debug/RPG.log
@@ -1,13 +1,0 @@
-﻿  Character.cpp
-D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\RPG\Character.cpp(195,6): warning C4101: 'Attack': unreferenced local variable
-D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\RPG\Character.cpp(196,6): warning C4101: 'Defence': unreferenced local variable
-D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\RPG\Character.cpp(197,6): warning C4101: 'Speed': unreferenced local variable
-  Combat.cpp
-  game.cpp
-  Inventory.cpp
-  item.cpp
-  main.cpp
-D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\RPG\Character.cpp(198,6): warning C4101: 'HP': unreferenced local variable
-  Generating Code...
-D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\RPG\main.cpp(7,12): warning C4244: 'argument': conversion from 'time_t' to 'unsigned int', possible loss of data
-  RPG.vcxproj -> D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\x64\Debug\RPG.exe
