--- RPG/RPG/x64/Debug/RPG.tlog/RPG.lastbuildstate
@@ -1,2 +1,0 @@
-#TargetFrameworkVersion=v4.0:PlatformToolSet=v142:EnableManagedIncrementalBuild=false:VCToolArchitecture=Native32Bit:WindowsTargetPlatformVersion=10.0
-Debug|x64|D:\All of My folders\Assignment\C++\Typing-based-Games\RPG\|
