+++ RPG/item.cpp
@@ -1,0 +1,12 @@
+#include "item.h"
+item::item()
+{
+	this->Name = "";
+	this->Buyprice = 0;
+	this->Sellprice = 0;
+}
+
+item::~item()
+{
+	//dtor
+}
